requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
filelock==3.18.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
python-whois==0.9.5
pytz==2025.2
requests==2.32.3
requests-file==2.1.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.6.0
tldextract==5.2.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
xgboost==3.0.0
